optimised data types used

+++ v21_gfp_exact_sched_test/2processors/scheduler.cpp
@@ -6,7 +6,7 @@
 
 
 // Global FP (priority ordering given by indices, index 1 == highest priority)
-bool prefer(const state& s, uint_fast8_t i, uint_fast8_t k) {
+bool prefer(const state& s, unsigned short i, unsigned short k) {
     if(s.c[i]==0 && s.c[k]==0) return (i < k); 		// no active job from i or k -- use default ordering
     if(s.c[i]==0 || s.c[k]==0) return s.c[i] > 0; 	// only 1 active -- prefer that one
     return (i < k); 								// FP rule (no ties are possible)
@@ -14,12 +14,12 @@ bool prefer(const state& s, uint_fast8_t i, uint_fast8_t k) {
 
 
 
-void sortTasksByPriorities(const state& s, uint_fast8_t* perm) {
+void sortTasksByPriorities(const state& s, unsigned short* perm) {
     
-    for (uint_fast8_t i = 0; i < s.n; i++) perm[i] = i;
+    for (unsigned short i = 0; i < s.n; i++) perm[i] = i;
     
-    for (uint_fast8_t i = 0; i < s.n; i++)
-        for (uint_fast8_t k = i+1; k < s.n; k++)
+    for (unsigned short i = 0; i < s.n; i++)
+        for (unsigned short k = i+1; k < s.n; k++)
             if (!prefer(s, perm[i], perm[k]))
                 swap(perm[i], perm[k]);
 }
